web/requirements.txt
django
mysqlclient
pycryptodome
gunicorn
raven
django-enumfield-ng
django-allauth
django-mailgun